attestation: Add new field to the Device Trust proto
This change adds a new new device_trust_signals_json string field to
the proto.This field is needed for the change in signals collection.
Now the signals will be collected in a dictionary and set as a JSON
string on the key info.

Design-doc:
https://docs.google.com/document/d/1OSYYezFB6iuEcFfr3fiVo00Sv8lgiLgvTINB9GKBFyw/

BUG=b:230646982
TEST=CQ

Change-Id: Ie28b19f6ce036b8e259d624286bc87bc46a4f512
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3633088
Reviewed-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Martin Rodriguez <rodmartin@google.com>
Reviewed-by: Martin Rodriguez <rodmartin@google.com>
Tested-by: Martin Rodriguez <rodmartin@google.com>