fusebox: add daemon FUSE op Write
Once a file is opened for writing, Kernel FUSE can send Write ops to
write a data buffer to the open file.

If the file has a local file descriptor, call WriteFileDescriptor to
locally write the buffer to the file. Otherwise, send a kWriteMethod
request to the server to write the buffer using server's file stream
writer service [1].

[1] Note for an MTP file, the data buffer is written to the server's
managed snapshot of the MTP file. A server kCloseMethod call will be
needed to copy to server snapshot to the MTP device (next CL).

BUG=chromium:1244007
TEST=CQ

Change-Id: I2d0897145cfa4130bb26893193cd7043b026a2e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3634977
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>