cryptohome: Generating OnboardingMetadata for crypto recovery.
This CL adds generation of OnboardingMetadata on a HW platform and
creation of recovery_id that identifies the cryptohome to be recovered.
The secrets used to generate a new recovery_id (seed and increment)
are stored in cryptohome and used on enrollment/re-generation only while
the actual onboarding metadata are part of the recovery blob (and hence
must be kept outside of the cryptohome).

BUG=b:192786922
TEST=cros_run_unit_tests --board=${BOARD} --packages cryptohome

Change-Id: Ib3d458b9b56d52fe6abfae5e10ed90cc6eaa6b17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3634809
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Tested-by: Maciek Slusarczyk <mslus@chromium.org>