power: Account for large time jumps for ChargeHistory
This utilizes base::TimeTicks to account for large time jumps in
base::Time. This can happen if the RTC loses state due to an empty
battery, which can cause the system time to initially be days or more
behind at boot. base::TimeTicks is monotomically increasing, and track
the time since boot. We utilize this to calculate a second duration. If
the difference between this duration and the base::Time calculated one
is too large, we adjust the plug in time for the AC charger (along with
the rest of the stats).

To allow unittesting of this, the power_manager::Clock is used, which
allows setting the wall time to different values.

BUG=chromium:1243163
TEST=manual and unittests

Change-Id: I8b151ade870cda6bd50b32562276778b1c02bb1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3633871
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>