hiberman: Really honor platform mode flag
First, fix a bug in snapdev, where for the PLATFORM_SUPPORT ioctl we
were passing a pointer to an int where we should have just been passing
the int itself. This was causing the kernel to see a non-zero value, and
always enabling platform support. This explains why regardless of the
command line flag we were going to S4 on machines that supported it.

With this fixed, it becomes obvious that the POWER_OFF ioctl doesn't do
anything if platform mode is not enabled. It simply returns. So, if we
want non-S4 behavior, we have to go to S5 ourselves. Add the libc call
to reboot() to power down in non-platform cases. This will be necessary
for dogfooding since S4 on Brya doesn't work without additional firmware
fixes.

BUG=b:204563517
TEST=hibernate on brya

Change-Id: I0b3e0e537fdd8b17d9a0aa1f6c62382ddbb2f33f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3635768
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>