chromeos-config: Add AC/DC support for resourced config
We need the ability to set different governor preferences depending
on the power source for the device. This change nests the existing
preferences under an ac/dc message. There is currently no one using
these messages, so it's safe to change them.

BUG=b:221091499
TEST=emerge-guybrush chromeos-config and verify new resourced properties
are populated.

Change-Id: Ied0ca5b24bd7acf6a8fbbcc0dfa706b53353da0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3639308
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>