cryptohome: Prevent UAF in AuthSession callbacks
Replace base::Unretained(this) with weak pointers in the callbacks bound
to AuthSession instances, in order to prevent even possibility of
use-after-free errors in case the callback outlives and gets fired after
AuthSession destruction.

The dangerous scenario was becoming increasingly possible due to the
integration of asynchronous auth blocks, as their lifetime is bounded to
the AuthBlockUtility singleton, as opposed to the typically short-living
auth sessions.

BUG=none
TEST=none

Change-Id: I315b721b1d9a790a519215d19bd2c8bb3331a096
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3641846
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Maksim Ivanov <emaxx@chromium.org>