cryptohome: Rename the USS file
Store UserSecretStash in the file called "uss.0" instead of just "uss".

This anticipates that we'll likely implement in the future a scheme with
storing multiple last versions of the file, so that it's possible to
revert to the last readable version in case of FS corruptions or losses.

Note that this commit, while non-backwards-compatible, is fine as the
USS is not launched yet.

BUG=b:232299885
TEST=cros_workon_make --board=$BOARD cryptohome

Change-Id: Ibf513849d3b0487b3b4cdc95c6f3d3ec611f3b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3646287
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>