arc: Qualify moves for LLVM roll
Currently, these moves are unqualified. These will produce a fatal
error with newer versions of LLVM. The toolchain team is working
to qualify all moves for the new compiler roll. The failing error is
`-Wunqualified-std-cast-call`.

These are the only known instances in platform2 code.

This change is required for the LLVM roll to r458507.

BUG=b:231351802
TEST=CQ
TEST=tast run arc.VTSKeymaster.vm  # On Hatch DUT

Change-Id: I3d058f87cf77d5dfc442847a0ee5c2e7dec9f9fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3647321
Reviewed-by: Edman Anjos <edman@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>