cryptohome: Make challenge_credential_info an optional parameter
ChallengeCredentialInfo is handled in
KeysetManagement::AddInitialKeyset() only if the key type is challenge
credential. Therefore there isn't any dangerous access within that
function. In the calling function this parameter used to be passed with
the Credentials object. Thus, if the Credential object didn't
instantiate |challenge_credential_info_| it was not a problem since it
was only accessed for challenge credentials case.
However, with combined AddKeyset flows for challenge
credential key type and other key types on AuthSession, passing the input
challenge credential info parameter (without Credentials object)
becoming an issue. The correct way of passing this parameter would be
making it optional since it is used only for specific challenge
credential case.

BUG=b:223221875
TEST=Unit tests and hwsec.* tast tests pass.

Change-Id: I4ad10705792cbae6dd16ac9acde693c04790fc05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3648148
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>