shill: metrics: define new types for enum and histogram metrics
This patch adds in Metrics two helper structs EnumMetric and
HistogramMetric to organize better the large amount of constant values
defined in metrics.h and to streamline the definition and usage of enum
and histogram metrics.

These two structs are templated with one additional "naming" parameter
that is used to standardize the process of splitting shill shill metrics
by technology with the name pattern "Network.Shill.$tech.$metric". The
template argument is essentially used as a dispatch argument to select
the correct wrapper aroudn SendToUMA/SendEnumToUMA. A new type can be
defined to specialize a SendToUMA function for every new naming pattern
that might be needed.

This patch only defines the new types. Usage is added in follow-up
patches.

BUG=b:172212984
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ic61d98e4d1e2971f079d9ec55db651ea630c58cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3655727
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>