hiberman: Create resume-init command
Create the resume-init command that can be called from the init script.
It will check the hibernate cookie, and if set, create dm-snapshots for
each of the known RW logical volumes (currently a hardcoded list).

Note: This is also the CL where we officially take a hard dependency on
LVM.

BUG=b:232806877
TEST=run hiberman resume-init to a test LV (rather than the real LVs,
which are in use during normal system operation)

Cq-Depend: chromium:3700916
Change-Id: I3562f1474fd4144d03b1a05d80b92e16274a9dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3699365
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>