hiberman: Expand cookie from bool to enum
Allow the hibernate cookie to store a couple of different values, rather
than a single boolean. This will allow us to indicate that a resume was
started but not completed, or that an abort (merging dm-snapshots back
to disk) was in progress during a power interruption.

For now, the actual usage of the cookie isn't changed (we just set it to
ResumeReady or NoResume). This will get expanded in subsequent changes.

BUG=b:232806877
TEST=cargo build, play with hiberman cookie command

Change-Id: Ice01bc7c3f93f96b8ebc0ad9a5a489718ab14924
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3700910
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>