cryptohome: Add VaultKeyset support to AddAuthFactor
Reuse existing AuthSession private methods to create KeyBlobs, encrypt
and persist a new VaultKeyset.

For initial user case, on session start a file system key is generated
and the mount operation is carried out with this key. If USS experiment
is enabled USS paylod is generated from that |file_system_key_| and
|auth_session.user_secret_stash_| is populated. AddAuthFactor()
function checks whether |user_secret_stash_| is populated and add the
initial AuthFactor with USS. Otherwise, if |user_secret_stash_| is null
AddAuthFactor() creates initial VaultKeyset by calling
AddInitialKeyset() with the same |file_system_key_|.

For a returning user, if USS experiment is enabled and there are no
VaultKeysets on the disk user is authenticated with USS and
|user_secret_stash_| is populated during this authentication.
AddAuthFactor() checks |user_secret_stash_| and take
AddAuthFactorViaUSS() path. If USS experiment is not enabled or there
are VaultKeysets on the disk authentication is done with VaultKeyset.

In this case |user_secret_stash_| is not populated and
AddAuthFactor() takes AddAuthFactorViaVaultKeyset() path.

BUG=b:223212595
TEST=FEATURES=test emerge-puff cryptohome
TEST=tast run 127.0.0.1:2022 cryptohome.*
TEST=via cryptohome CLI
 - cryptohome --action=start_auth_session
 - cryptohome --action=authenticate_auth_factor --password
 - cryptohome --action=add_auth_factor --pin
 - cryptohome --action=authenticate_auth_factor --pin

Change-Id: I2c50b1e35d88f5c012a20ed14b24492e4cb00d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3587151
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>