shill: Expose wifi signal strength dBm value for reporting.
We currently report wifi signal strength dBm value in chrome by
converting the 0-100 signal strength using the inverse of
WiFiService::SignalToStrength() this can lead to wrong conversion
if the conversion logic was out of sync. So this change exposes
the dBm value to be used directly instead.

BUG=b:232966507
TEST=FEATURES=test emerge-${BOARD} chromeos-base/shill

Change-Id: Iff7815c15cd9165cada9cd7d4aebcc2ed26222a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3652883
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Ahmed Nasr <anasr@google.com>
Commit-Queue: Ahmed Nasr <anasr@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>