shill: Add EphemeralPriority property in Service
So that we can sort Ethernet services in test.

The existing Priority property is not suitable here because the profile
interaction when there are multiple Ethernet services can be tricky.

In the current implementation, we always make the generic Ethernet
service ("ethernet_any") pointing to the first Ethernet service (see `
EthernetProvider::RefreshGenericEthernetService()`), and this logic is
executed once services order changes (in `Manager::SortServicesTask()`),
and thus it will also happen when we set Priority property on an
Ethernet service which is not the one with the highest priority. One
result of this is, Manager::MatchProfileWithService() will be called on
this service which triggers a Service::Load() from the "ethernet_any",
and then the |priority_| of this service will be reset to 0 if
"ethernet_any" does not have a value for that. The services order will
not be effected by this reload since it does not trigger a sort, but it
results that this order is not stable.

Thus we need a priority value that is not persisted into profile.

BUG=b:192436642
TEST=unit tests

Change-Id: I4b6bf71c6684c92fa784556e4caeec30e9363adb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3653642
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>