init: Mount efivarfs
Update chromeos_startup.sh to mount efivarfs. This provides access to
EFI variables for boards that use efivarfs, rather than efivars.

Mounting is gated on the presence of "efivarfs" in /proc/filesystems,
which will only show up once the efivarfs module loads. On non-efi
systems or efi systems using efivars instead, it won't be present.

Testing needs to show that this works for boards with efivarfs in the
kernel, and that this doesn't interfere for boards without efivarfs.

Specific test:
```
$ cros flash file:///images/betty.bin xbuddy://remote/betty/latest
$ runvm --efi -- /images/betty.bin
$ ssh vm ls /sys/firmware/efi/efivars/
> no output
$ ssh vm ls /sys/firmware/efi/vars/
> expected list of Boot* and other efi variables
```

with crrev/c/3630739 the reven board uses efivarfs:
```
$ runvm --efi -- src/build/reven/latest/chromiumos_test_image.bin
$ ssh vm ls /sys/firmware/efi/efivars/
> expected list of Boot* and other efi variables, same as above plus
>   MokList-<GUID>/MokListRT-<GUID> and SbatLevelRT-<GUID>
$ ssh vm ls /sys/firmware/efi/vars/
> ls: cannot access '/sys/firmware/efi/vars': No such file or directory
```

BUG=b:190430369
TEST=See comment

Cq-Depend: chromium:3630739
Change-Id: I68be1b54ff9a814ac4ad311a6839715f6d2d2d63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3635797
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>