oobe_config: Add tgkill to seccomp
oobe_config_restore is crashing and returning sig 31 error. This
error is logged because tgkill has been triggered and it is not
included in seccomp.

Add tgkill to seccomp. Although this does not fix the cause of the
crash, logging sig 31 interferes in the debugging of the cause.
Therefore, this will allow to log the reason of the crash instead
when a crash of this type happens.

Add tgkill also to oobe_config_save in case there is a crash in
the future.

BUG=b:233854839
TEST=sig 31 error is not returned when oobe_config_restore crashes.
sig 31 is not returned when oobe_config_save crashes.
Both scenarios have been tested in arm, arm64, and amd64 devices.

Change-Id: Ia3662c998fc5408b28af589586be305afaa6c579
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3654106
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>