shill: support WiFi session tags for metrics
Initial implementation of the WiFi Session Tags.

We use those to "tag" WiFi structured metrics events such as
|WiFiConnectionAttempt|, |WiFiConnectionAttemptResult| and
|WiFiConnectionEnd|. That way, the server-side pipeline has the ability
to know which "disconnection" event corresponds to which "connection
attempt" for example. It's particularly useful if some events are
missing (bug in the client-side implementation, crashes, data loss
between the client and the UMA server, throttling, ...), otherwise the
server could try to match the wrong "connection attempt" and "connection
attempt result" when rebuilding a session.

See more details at go/cros-wifi-metrics-session-tag-dd

BUG=b:226145383
BUG=b:225737253
TEST=cros-workon-${BOARD} start shill
TEST=emerge-${BOARD} shill
TEST=FEATURES=test emerge-${BOARD} shill
TEST=cros deploy ${DUT} shill
TEST=Enable/disable the WiFi adapter 3 times and observe the sequence of
session tags in the log. The events that are emitted have the correct
tag, but the disconnection event is not emitted, to be fixed in a
follow-up CL.
TEST=Connect to an AP, then manually connect to another AP, observe the
correct sequence of session tags in the log
TEST=restart shill SHILL_LOG_SCOPES=metrics+wifi SHILL_LOG_LEVEL=-4
TEST=tast -verbose run -var="router=${ROUTER}" -var="pcap=${PCAP}"
${DUT} "wifi.SimpleConnect.*"
TEST=tast -verbose run -var="router=${ROUTER}" -var="pcap=${PCAP}"
${DUT} "wifi.RoamAPGone"
TEST=tast -verbose run -var="router=${ROUTER}" -var="pcap=${PCAP}"
${DUT} "wifi.DisconnectReason.ap_gone"
TEST=tast -verbose run -var="router=${ROUTER}" -var="pcap=${PCAP}"
${DUT} "wifi.DisconnectReason.deauth_client"

Change-Id: I28f3489079eb4af0fc535e5d50574e3b49ab5ca6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3659255
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>