u2fd: Fix pubkey format used in attestation
In a previous refactor we hid the raw public key to cbor format
transformation in the u2f processor because it's up to processors to
decide the raw format, and WebAuthn only cares about the cbor format.
But actually we're using the "fido-u2f" attestation format that expects
a raw ECC pub key format in the attestation statement, so in practice
both cbor and raw format public keys have to be returned. However, raw
format public keys only need to be returned in fido-u2f compat
processors. Modify the interface and related unittests, and fix the bug
that passed cbor pub key format to attestation statement.

BUG=b:226453823,b:233717384
TEST=u2fd unittests
TEST=manual test that enterprise attestation is successful.

Disallow-Recycled-Builds: test-failures
Change-Id: I2ecd3a3afea0259ac133cf3b5495803a1b982bd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3653722
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>