chargesplash: Implement frecon abstraction
Implement a wrapper class for managing a Frecon process.

The object allows for writing to multiple output files.  The purpose
of this is so that, for automated testing, output can be written
either in addition to or instead of a real Frecon terminal.

BUG=b:228370390
TEST=cros_workon_make --board=ghost chargesplash --test

Change-Id: I975f6f269da42d450d9dc176bf66edeca8af8278
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3659253
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>