shill: don't fail unit tests with extra logs
Some unit tests are checking the log messages that are added to the log
by the methods they're testing. In that case the test also needs a
"catch all" call to ensure that the test doesn't fail when the internal
implementation changes and adds another log message on top of the
expected log message.

BUG=None
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: Ic76cf4ccec5273d9b48d49f4a958b95328db7374
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3659254
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>