dns-proxy: Resolver handle all queries results
As part of implementing probing, the resolver needs to know
about all of the queries result in order to invalidate a
name server / DoH provider. This patch moves the responsibility
for handling multiple queries result from the clients to the
resolver.

Note that in the follow-up patches, the clients responsibility
will be reduced even more such that it is only responsible for a
single query to a name server / DoH provider. All the logic will
be moved to the resolver.

BUG=b:228247828
TEST=dig google.com;
     works with multiple name servers;

Change-Id: I7d98ea3a543aa419094655007e703a1343bab3aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3676703
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>