hiberman: Implement dm-snapshot merging
Add functionality to the VolumeManager that enables it to merge
dm-snapshots set up by the resume-init command. This mechanism will be
used by the "abort" path, where we set up dm-snapshots on top of RW
stateful regions, only to decide later to cancel the resume from
hibernate.

Currently this is wired up to the resume path, so that if
ResumeConductor::resume() ever returns, we'll do this sync
operation.

I've tested this by manually creating a new "evantest" LV, then editing
the code to manage just this test LV. With this, I can run hiberman
resume-init, mount the "evantest-rw", write to it, then run this new
code to exercise the abort path.

BUG=b:233103309
TEST=Manually try on a test LV.

Change-Id: I2bc5224550a2ae4df68dc04b76ccb283f664d4eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3700911
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>