minios: Add progress bar to finalization screen
The finalization screen is the longest step in the NBR recovery
process. No visual cues are currently provided to the user which
can lead to the perception that the system is frozen. To avoid
the wrong perception we now display an indeterminate progress
bar during the finalization step.

BUG=b:217629874
TEST=FEATURES=test emerge-$B minios

Change-Id: Ic40fda30e37b61699169f2ecf52a9a89f0884017
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3665785
Tested-by: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Yuanpeng Ni‎ <yuanpengni@chromium.org>