shill: Add session tags to WiFi structured metrics
The "session tag" feature appears to be stable enough to start adding
the tags to the WiFi structured metrics events. A few edge cases remain:
- "disconnect" arrives before "connection attempt result"
- roaming, in particular failure to connect to the new BSSID

In those edge cases the logic works as expected and the session tag is
set to invalid (0) for the problematic event, which will let the
server-side pipeline discard those events and the corresponding
sessions, similarly to cases where some events would be lost.

Overall it's better to enable the tag compared to the current
situation where the tag is always 0 and the server-side pipeline
actually gets less information.

BUG=b:226145383
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: Ic4f8466185ab5c78375c8f3c928d410da24efdd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3661559
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: David Ruth <druth@google.com>