modemfwd: remove noisy helper seccomp policy log
We log every time when a helper is run with a seccomp policy to confirm
the sandbox is being called correctly. Now that the feature has had time
to stabilize, this log should be removed, especially since we log a
warning when this argument is missing.

BUG=None
TEST=built modemfwd with change.

Change-Id: I2a0e310be9ff140acca28dba7d027c806c6c3d4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3665316
Auto-Submit: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>