dns-proxy: Resolver handles clients' concurrent queries
Moves the logic of querying multiple name servers and
DoH providers to the resolver. As the clients are no
longer handling the results of multiple concurrent queries,
it would be cleaner to let resolver handle the concurrent
queries. This would allow the clients to only have a single
responsibility of querying a single name server / DoH provider.

For ares, the behavior for starting the concurrent queries
is slightly different resulting in a small performance drop.
Before, the concurrent queries are done:
1. create socket1, send query1,
2. create socket2, send query2,
3. process all sockets (e.g. epoll_ctl add).
With the patch, the concurrent queries are done:
1. create socket1, send query1,
2. process socket1 (e.g. epoll_ctl add),
3. create socket2, send query2,
4. process socket2 (e.g. epoll_ctl add).
The processing in step (2) and (4) above takes around 0.3 ms
on atlas. So in the new behavior, query2 will be executed
0.3 ms later, query3 will be executed 0.6 ms later, etc.
The number looks insignificant in this case as queries on
average takes ~20 ms (data from UMA metrics).

BUG=b:228247828
TEST=dig google.com;
     see that all name servers are used;
TEST=dig google.com;
     see that all DoH providers are used;
TEST=test that performance change is insignificant;
     on atlas;

Change-Id: I242614a1ed2e1025ebe13480b0d0c75710aa02e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3676707
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>