runtime_probe: Add gpu probe function
Add gpu probe function to probe gpu components. The sandbox config for
arm platforms are omitted. It will be added once we support it on arm.
Before that, just let the function raise error so the users know that it
is not yet supported.

BUG=b:213117002
TEST=FEATURES=test emerge-zork runtime_probe
TEST=FEATURES=test emerge-reven runtime_probe
TEST=Test on dut

Change-Id: I827af72c2035a6659e35b5cff3e70daab3992775
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3667178
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>
Tested-by: Chung-Sheng Wu <chungsheng@google.com>