fpstudy: Format all python in analysis-tool
We need to reset the formatting once, so let's just rip off the
band-aid now, rather than scattered in with later changes.

The annoying part is that it is combining imports of local modules
with external modules. We probably need to do a restructuring to
make it clear that those imports are local libraries.

cros format analysis-tool/*.py analysis-tool/benchmarks/*.py
BUG=b:339298572
TEST=None

Change-Id: I2a6fc5e5b663d89d2e646bc03755d9a9feb64d69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5550410
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>