cryptohome: Add missing reset seed to AuthSession's VaultKeyset
When a user with a very old password VaultKeyset adds a PIN
VaultKeyset, the new PIN VaultKeyset won't have a reset seed. So the
wrong attempt locks cannot be reset. To avoid this scenario, when
AuthSession::AddCredentials() is called we should check whether the
existing VaultKeyset has a reset seed and add if missing. This added
reset seed will be transferred to all new added keysets during the
KeysetManagement::AddKeyset() call and will be common among all the
VaultKeysets.

BUG=b:223222440
TEST=Unit tests, cryptohome.* and hwsec.* tast tests pass

Change-Id: I16f55e030615096daadfed258c0642e12b9e9b0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3652485
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>