nnapi: add generated (modified) AIDL code and stubs
This CL adds the generated code from the AOSP AIDL definitions that
required any changes. It also adds the AIDL generated code into the
build.

It adds in the parts of the Binder framework that is needed for
compilation of all the AIDL generated binding code, modified where
necessary for compatability with ChromeOS and / or to not pull in deeper
parts of libbinder.

Finally, it adds in a series of C stub calls for the Binder framework
which should never get called, but are required for successful linking.

BUG=b:234073079
TEST=FEATURES=test emerge-brya nnapi

Cq-Depend: chromium:3670957
Change-Id: Id34b820de041f7fd683e03985903e23dbc4be6a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3671002
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Jack Shen <jackshen@google.com>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>