vm_tools: Retry inside ContainerReady
We're seeing a lot of container start timeouts where LXD reckons it's
booted a container, but Chrome never gets a response. My hypothesis is
that there's a gap between a container being booted enough for
StartContainer to report started, and it being running enough for
GetContainerInfo to consider it running.
If Cicerone is too quick (or the container is too slow) then it'll try
getting container details such as IP address before it's ready, fail,
and drop the event.

Try working around this by retrying after a minute to see if it helps.
We see lots of test failures so if this is indeed the cause we should
see an improvement quite quickly and can spend more time making it
nicer, or if it doesn't then we don't need this commit at all.

BUG=b:233831053
TEST=CQ, then will check stainless in a bit

Change-Id: Ied7fa86cb54231afde28dbaaf35f89a208487b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3671006
Commit-Queue: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Tested-by: David Munro <davidmunro@google.com>
Reviewed-by: Alexey Shpakov <ashpakov@google.com>