cryptohome: Add Reap*Error() functions for disposing errors
The Reap*Error() series of functions are designed so that there's a
centralized way to dispose of errors that does not propagate all the way
up the stack, back to the DBus call. This is so that those errors will
not be silent.

BUG=b:229931701
TEST=Builds and tested with CLs later down the chain

Change-Id: I74ac5c1137b0fed3fdb6694251faa0e3117f9266
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3678068
Commit-Queue: John L Chen <zuan@chromium.org>
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Yi Chou <yich@google.com>