shill: set disconnection type in WiFiConnectionEnd
Report the type of disconnections in the |WiFiConnectionEnd| event. If
the disconnection comes from the AP, it's always unexpected. If it comes
from the station it's expected in cases where the service was expecting
a disconnection, for example when the user selects another SSID or when
the user manually disconnects.

BUG=b:225737253
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill
TEST=emerge-${BOARD} shill
TEST=cros deploy ${DUT} shill
TEST=tast run -var="router=${ROUTER}" -var="pcap=${PCAP}" ${DUT}
"wifi.RoamDbus" and observe in the log a "roaming" disconnection and an
"expected" disconnection
TEST=tast run -var="router=${ROUTER}" -var="pcap=${PCAP}" ${DUT}
"wifi.RoamAPGone" and observe in the log a "unexpected by STA"
disconnection and an "expected" disconnection
TEST=Connect to SSID1, then manually select SSID2 and observe in the
log an "expected" disconnection.
TEST=Connect to SSID1, then manually disconnect in the UI and observe in
the log an "expected" disconnection.

Change-Id: Ie6b014e3477a0939f03845a57ed1cf46be28c0c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3673603
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>