dns-proxy: Fix ares TCP query
Ares client TCP queries were failing because of a bug where
it uses WatchReadable as opposed to WatchWritable. This makes
the TCP query to be run only when the connection is closed by
the other end of the connection.

The TCP query is triggered when a response is larger than the
size of DNS max packet size (512 bytes), which sometimes happen
even without EDNS.

Alongside fixing the bug, this patch also update the behavior
such that any TCP query will be proxied with a TCP query, the
same for UDP. The client is expected to handle any protocol
change, etc.

BUG=chromium:1329388
TEST=dig +tcp TXT google.com; # Over 512 bytes.
     works and goes through TCP;
TEST=dig TXT google.com; # Over 512 bytes.
     works and goes through UDP;

Change-Id: I561bf4e6cbe86e390edd01dd408a971cdf3dc7cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3676634
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>