vm_tools: init: Don't call chown in script stanza.
Although we were executing `chown` against virtio-vhost-user proxy
devices on ManaTEE in the script stanza, we are not allowed to call
external commands like `chown` in the script stanza.
Otherwise, upstart won't track concierge's PID and won't kill the daemon
when a user sign out.

So, this CL moves a logic for chown to the pre-start stanza.

BUG=b:233572447
TEST=login and check if `initctl status concierge` shows PID

Change-Id: Ie663d8b0b3b46fb1ab16fd1c88449b5b7ce12409
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3680038
Tested-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>